메모리 및 pte 에러 수정, 패치 버그 수정, faf 및 vfork_done 관련 수정

See merge request iitp-sds/centos_kerrighed_kernel!3

+++ kerrighed/mm/mm_server.h
@@ -30,7 +30,7 @@ typedef struct mm_mmap_msg {
 	};
 	union {
 		unsigned long new_len;
-		unsigned int vm_flags;
+		unsigned long long vm_flags;
 		unsigned long prot;
 	};
 	union {
